Rename NotesMindMapView -> NotesMindMap
