+++ src/components/NotesMindMap/NotesMindMapStyles.js
