+++ src/reducers.js
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux';
 
 import { appReducer } from 'components/App/AppReducer'
 import { notesReducer } from 'components/NotesPage/NotesPageReducer';
-import { notesMindMapReducer } from 'components/NotesMindMapView/NotesMindMapViewReducer';
+import { notesMindMapReducer } from 'components/NotesMindMap/NotesMindMapReducer';
 import { searchPanelReducer } from 'components/SearchPanel/SearchPanelReducer';
 
 export const reducers = combineReducers({
